Bug#34641458 ClusterJ test case crash at end
Problem:

When a ClusterJ program exits the NdbAPI context is cleaned up by a call
to ndb_end() triggered by the destructor for a JTie_NdbInit object with
static storage duration as part of the program exit (see Bug#27869124).

This call races with destruction of other static storage duration such
as proc_set_array in NdbLockCpuUtil.cpp which is a std::unique_ptr.

The destruction order between these objects are not defined.

The ndb_end() function calls indirectly NdbLockCpu_End() function which
accesses the proc_set_array.

What happen if proc_set_array is destroyed before the JTie_NdbInit
object is undefined it may work or not.

This race was introduced in MySQL Cluster 8.0.30 where proc_set_array
was changed from being a raw pointer to an unique pointer.

Solution:

Make proc_set_array a raw pointer again to avoid race.

Change-Id: I3cc706d38d125a95ccde72bafee9d9253ab431cb