WL#15130 Socket-level TLS patch #14: testSecureSocket
This patch introduces the test program testSecureSocket-t,
with 2-thread send, send/recv, readline, and writev tests.
With the -s option, testSecureSocket runs TCP and TLS echo
servers. With the -c option, it can run any of the tests.
Each test has a two-thread design where one thread sends
data to the echo server and the other thread receives the
replies.

With no options, this runs as a TAP unit test that starts
plain and TLS servers on ports 3400 and 3401 and runs the
first 12 tests.

Change-Id: I23060facbfe517ef33186de3be24e6916b3b56ac