Bug#34682586: MySQL server crash, segmentation fault: count_field_types
A grouped query crashed on the second execution when using the
hypergraph optimizer.

The crash happened because the hypergraph optimizer added a hidden
item to Query_block::fields in the first execution, and failed to
remove it before the next execution. The added item itself was
destroyed. In the second execution, the dangling pointer to the
destroyed item was accessed, and the server exited because of a
segmentation fault.

The hidden item was added in order to ensure that all columns in the
ordering were available in a materialization node below the sort node.
However, adding this particular column turned out to be unnecessary,
because it came from an access path that was not part of the
materialized subtree in the first place. As it turns out, none of the
queries in the test suite need to have a hidden element added here,
even though many of them get one. And it's not entirely clear if there
are cases where we need to add one.

Since this piece of code has been a source to other bugs before as
well, and it does not seem to fix any known problem, this patch
removes it and instead adds an assertion that verifies that the sort
node does not access any column that is not available to it.

Change-Id: I63a0b1def683472943fa2bcca5cbdb8d8fc140ad