Bug# 33630199 - mysqldump make a non-consistent backup with --single-transaction option
Problem
-------
This problem arise when we have a server which has GTIDs enabled, and
there exists few on going transactions in the background or in other sessions.
Basically those kind of transactions which will cause GTIDs to increase.

When executing mysqldump to dump data of this server, with options
--single-transaction and --set-gtid-purged=ON, the dump file generated
has inconsistent data. When examining the dump file, we'll notice that
the dump file has the GTIDs executed for those transactions which were
running in the background, but it does not have the corresponding data
for those GTIDs.

Analysis
--------
When --single-transaction option is specified, a snapshot of the
database at that instant is taken. But the transactions which were
happening in the background are still causing GTIDs to increase.

After a transaction over a database snapshot is started, mysqldump
disables the binlog, unlocks the table, proceeds with the dump, once
dump is over, it fetches the GTIDs. Since GTIDs are being fetched
after the tables were unlocked, GTIDs would have been increased
during the period when dump was running, since many transactions
which were running in the background were completed, resulting in
increment of GTIDs.

Solution
--------
Current flow for mysqldump is as follows:
	1. start transaction
	2. disable binlog
	3. unlock tables
	4. start the dump
	5. dump finish
	6. fetch GTID_EXECUTED, and print it

Here, we see that when the GTIDs are being fetched, tables were
unlocked already, and dump was over too.

Fix does the following:

  1. mysqldump will check if --single-transaction is enabled, if it
  is enabled, mysqldump will perform FTWRL by calling the function
  do_flush_tables_read_lock(mysql);
  This will perform a read lock on the tables.

  2. "my_bool single_trans_with_gtid_enabled" will check if the
  server has GTIDs enabled, and mysqldump was started with the
  option --single-transaction, and --set-gtid-purged is NOT off.

  3. Locks acquired, start the transaction, get the snapshot of the
  database. Now it can fetch the value for GTID_EXECUTED, and store
  it in the variable "MYSQL_RES *gtid_executed". This variable is
  initialized to NULL in the beginning of main().

  ** MYSQL_RES doesn't have a life of its own, it's only valid as
     long as the MYSQL that produced it is in a matching state.
     Hence, it is not preferred to pass the same variable to other
     functions.

     New struct MY_GTID_ARRAY has been introduced, which can hold
     the result of `SELECT @@GLOBAL.GTID_EXECUTED`. It has 2 members,
     one will tell the number of rows, and other one is just an array
     of strings, to store the GTIDs.

     The `if(single_trans_with_gtid_enabled)` block will basically
     perform a deep copy of gtid_executed into the MY_GTID_ARRAY,
     and store result of `SELECT @@GLOBAL.GTID_EXECUTED` into
     gtid_executed.

Thus, this MY_GTID_ARRAY will be passed over to other functions.

Till now, we've both GTIDs executed, and the corresponding data in
the snapshot.

  4. The function responsible for fetching the values of GTIDs
  is

    > add_set_gtid_purged(MYSQL *mysql_con)

  which is being called by the function

    > process_set_gtid_purged(MYSQL* mysql_con, my_bool flag)

  process_set_gtid_purged now takes two more parameter, viz.
  MY_GTID_ARRAY **gtid_array, my_bool single_trans_with_gtid_enabled
  ,and so does the add_set_gtid_purged.

    > process_set_gtid_purged(MYSQL* mysql_con, my_bool flag,
			     MY_GTID_ARRAY **gtid_array,
                             my_bool single_trans_with_gtid_enabled)

    > add_set_gtid_purged(MYSQL *mysql_con,
                          MY_GTID_ARRAY **gtid_array,
                          my_bool single_trans_with_gtid_enabled)

So, basically, the GTIDs which we stored initially after the
FTWRL, are now being passed to this function. And inside of
the process_set_gtid_purged, the same set of GTIDs is being
passed to add_set_gtid_purged.

    - If mysqldump was NOT started with --single-transaction,
    then we did not store GTIDs in the beginning of the dump.

The only condition for this bug is to have a server with gtids
enabled, and a dump started with --single-transaction.
`single_trans_with_gtid_enabled` is keeping track of this condition.

If `single_trans_with_gtid_enabled` is false, then

    add_set_gtid_purged checks if it's false, if it is, it means that
    GTID_EXECUTED wasn't fetched earlier, in that case, run
    the query again to grab GTIDs, then proceed normally. This is
    the normal behaviour, that is, getting the GTIDs after dump.

    Depending on whether `single_trans_with_gtid_enabled` is true
    or false, add_set_gtid_purged will take care of the ways to
    fetch gtids accordingly.

  5. We've now both, snapshot of the data, along with their
  corresponding GTID set.

Earlier versions used to fetch GTIDs at the beginning of the dump,
but the sequence in which GTIDs are fetched has changed because of
the issue caused by Bug# 32843447. Changing the order permanently
in which GTIDs are fetched might trigger that bug again. This is
the reason that only for this particular case the order in which
GTIDs are fetched has changed.

With these changes, the original functionality remains untouched.

Added few more functions

  > static int get_gtid_mode
     This function checks if GTIDs are enabled on the server. And
     return integer result.

  > static my_bool get_gtids_executed
     This function will store the table GTID_EXECUTED in the
     parameter gtid_executed.

  > struct st_my_gtid_array (typedef to MY_GTID_ARRAY)
     A structure to store results for GTID_EXECUTED.

Added new warning in the function : process_set_gtid_purged

  >  Running mysqldump on a GTID enabled server, without
     --single-transaction or --master-data or --lock-all-tables
     will throw a warning about data dump might being inconsistent.

Change-Id: I90347497275fded6a1896b76c6f7a1b04b28fc2e