Bug# 33630199 : mysqldump make a non-consistent backup with                 --single-transaction option [mysql-trunk]
Problem:
--------
Dumping data using mysqldump with --single-transaction and
--set-gtid-purged=ON, while on the other session, INSERT
queries are being sent, the dump file generated has inconsistent
data. That is, the backup data has the gtids but not the
corresponding data.

Analysis:
---------
When --single-tranasction is used, a transaction is started, which
basically creates a snapshot of the data which is currently present
in the database. After that, the function process_set_gtid_purged
takes care of the gtids present in the system, fetching them, and
adding the GTID_PURGED to the dump file. There's a short time window
between these two things. If someone executes an INSERT statement
or other queries during this fraction of second, the data won't be
there in the dump since START TRANSACTION took a snapshot of the data
before this moment, but the GTID set will contain these extra transactions.
And the resultant dump file will have the old data with the new GTID
set resulting in inconsistent dump.

Fix:
----
1) The fix is to use FTWRL (FLUSH TABLES WITH READ LOCK). Here, the
   point is that FTWRL allows us to both start a transaction with
   consistent snapshot, and consistently read the value of
   gtid_executed for that snapshot.

2) If the dump was started with --single-transaction, we would
   just call do_flush_tables_read_lock, this would acquire read
   locks on the tables. The FLUSH TABLES statement present inside
   of this function would also cause an increase in GTID by one.

3) Once locks have been acquired, just proceed normally.

4) The same bug existed in mysql-5.7, but in that case, mysqldump
   takes the GTID_EXECUTED when data dump is over, and the tables
   are unlocked, but in case of mysql-8.0, the GTID_EXECUTED is
   stored before the tables were unlocked, so in this case, the
   fix was just to perform a FTWRL before fetching the GTID_EXECUTED.

Added new warning in the function : process_set_gtid_purged
  >  Running mysqldump on a GTID enabled server, without
     --single-transaction or --master-data or --lock-all-tables
     will throw a warning about data dump might being inconsistent.

Change-Id: If1e6f4b09ba398ade5e9bda20934db510b8871dc