BUG#34619134: Asynch replication with consistency mode BEFORE causing assert failure
The implementation of the BEFORE consistency of Group Replication,
includes a call to `Gtid_state::wait_for_gtid_set()`, that call is
preceded by a assert that ensures that the caller session does not
belong to a Replication applier.
Though that is a valid scenario, we can have a group, on which we
have `group_replication_consistency=BEFORE` set on the global scope,
to which we replicate from a outside server:

 |---|
 | O |
 |---|
   |
   |
   |      |---|   |---|  |---|
   \----->| P |   | S |  | S |
     ch1  |---|   |---|  |---|
            |       |      |
            \_______|______/

The assert is trying to protect the case on which a statement like
 `INSERT INTO t VALUES (WAIT_FOR_EXECUTED_GTID_SET('...'))`
comes through a Replication applier session.
That case is already covered on
`Item_wait_for_executed_gtid_set::val_int()`, which will return a
error if executed through a Replication applier session.
Therefore, we remove the incorrect assert.

Change-Id: Ifd5104392d64b14ec9acd6dd1dc8383d2ae6dc0b