Bug#34637153: Performance of Hypergraph is lower than the old optimizer for point selects [5/7, mem_root_unordered_map, noclose]
CostingReceiver stores the plan candidates in a std::unordered_map.
This leads to dynamic memory allocation when adding candidates to it.
This patch replaces it with a mem_root_unordered_map, which allocates
memory on the execution MEM_ROOT instead. The AccessPath objects that
are stored in the map, and which consume more space than the map
itself, are already stored on the execution MEM_ROOT.

Microbenchmark results:

BM_FindBestQueryPlanPointSelect    1545 ns/iter -> 1471 ns/iter [+5.0%]

Change-Id: I38a14d8faefa9618e3b4ab9b51587e0f6aa31145