Bug#34637153: Performance of Hypergraph is lower than the old optimizer for point selects [6/7, sargable, noclose]
When analyzing equality predicates to see if they are sargable, some
time is spent checking that the arguments have compatible types so
that the predicate may be used for an index lookup. When the equality
predicate comes from a multiple equality, we already know that the
arguments have compatible types, since multiple equalities are
constructed only if the arguments have the exact same type.

We now skip the check for compatible types if the equality comes from
a multiple equality.

One bug in the multiple equality construction had to be fixed in order
to get this to work: A multiple equality was created for a string
column compared to a constant JSON expression, if the string column
had the same collation as the JSON expression (utf8mb4_bin). The
predicate would therefore be considered sargable, and a lookup was
performed on the index for the string column. This made the predicate
use string comparison semantics instead of JSON comparison semantics,
and wrong results were returned. This was fixed by adding an extra
check to check_simple_equality() to prevent creation of multiple
equalities mixing JSON and string types.

Microbenchmark results:

BM_FindBestQueryPlanPointSelect    1471 ns/iter -> 1432 ns/iter [+2.7%]

Change-Id: I21788c44668ef9c5d2ca2c7d7c6d0a7163be3c35