WL#15145 ndb binlog writeset trans dependencies [#1]
The shadow tables used by ndb_binlog thread when handling data events
have no knowledge about wheter the table is a foreign key parent in NDB,
this is mainly because the NDB dictionary extra metadata for parent
table is not updated when creating the child tables. The information
about being parent table is required in order to properly detect a
transaction conflict while writing changes to the binlog for changes of
parent key tables.

This is fixed by extending the ndb_binlog thread with a metadata cache
that maintains a list of tables in NDB which are foreign key parents.
This makes it possible to update the shadow table's knowledge about
being parent table while handling data events.

Change-Id: Ic551fde3e9460e3668d5aa1ac951ee3bec442cf5