BUG#34728079: Move error log message ER_GRP_RPL_SUPER_READ_ON to before set super_read_only
When Group Replication sets `super_read_only` it logs the messages
```
ER_GRP_RPL_SUPER_READ_ON
  eng "Setting super_read_only=ON."

ER_GRP_RPL_SUPER_READ_OFF
  eng "Setting super_read_only=OFF."
```
though despite the word "Setting" the log messages are logged after
the operation was attempted.
Which is confusing when the operation fails, the log will show:
```
2022-10-21T13:47:51.414696Z 32 [System] [MY-013587] [Repl] Plugin group_replication reported: 'Plugin 'group_replication' is starting.'
2022-10-21T13:47:51.414726Z 32 [Note] [MY-011716] [Repl] Plugin group_replication reported: 'Current debug options are: 'GCS_DEBUG_NONE'.'
2022-10-21T13:49:51.415752Z 32 [System] [MY-011565] [Repl] Plugin group_replication reported: 'Setting super_read_only=ON.'
2022-10-21T13:49:51.415785Z 32 [ERROR] [MY-011636] [Repl] Plugin group_replication reported: 'Could not enable the server read only mode and guarantee a safe recovery execution'
```
The operation to set `super_read_only=ON` has a 2 minutes timeout.
Note that there is no time gap between the attempt and the error.

The reader expects that the message is logged before doing the
operation, which would mean a log like:
```
022-10-21T13:47:51.414696Z 32 [System] [MY-013587] [Repl] Plugin group_replication reported: 'Plugin 'group_replication' is starting.'
2022-10-21T13:47:51.414726Z 32 [Note] [MY-011716] [Repl] Plugin group_replication reported: 'Current debug
options are: 'GCS_DEBUG_NONE'.'
2022-10-21T13:47:51.415752Z 32 [System] [MY-011565] [Repl] Plugin group_replication reported: 'Setting super_read_only=ON.'
2022-10-21T13:49:51.415785Z 32 [ERROR] [MY-011636] [Repl] Plugin group_replication reported: 'Could not enable the server read only mode and guarantee a safe recovery execution'
```
that is, a 2 minutes gap between the attempt and its error.

To match the log reader expectations, we now log the messages
before doing the operation.

Change-Id: Ieb9c7e99556950441294d0e5e18e16e5f3b702ec