Bug#34727088: Assertion `!OrderItemsReferenceUnavailableTables(path, tables)' failed.
A query hit an assertion in debug builds and returned wrong results in
release builds when running with the hypergraph optimizer, because it
tried to sort the results on a column that was not available.

The problem happened when doing a sort with duplicate removal for a
DISTINCT clause, and there was a streaming step somewhere below the
sort access path. The sort path sorted the rows on a column that was
not mentioned in the SELECT list, and the column was therefore not
materialized in the streaming step before the sort. The sort used the
column value that happened to be available from the base table scan,
instead of getting it from the temporary table, and this caused wrong
results.

The reason why the sort access path sorts on a column not in the
SELECT list, is that the hypergraph optimizer tries all available
sort-ahead orderings for satisfying DISTINCT in the hope that one of
them might also satisfy the ORDER BY clause that is processed just
after DISTINCT. It picks a sort-ahead ordering that satisfies the
ordering required for DISTINCT (via functional dependencies) even
though it doesn't reference the column in the SELECT list. This
ordering is picked only because it is seen first; not because it is in
any way better than sorting on the original column.

This patch changes how the sort for DISTINCT is chosen. Instead of
proposing all sort-ahead orderings that would satisfy DISTINCT, it now
proposes:

1) The sort-ahead ordering that is directly derived from the DISTINCT
clause in the query.

2) If there is an ORDER BY clause, all sort-ahead orderings that
satisfy both DISTINCT and ORDER BY.

This makes it prefer orderings that reference columns in the SELECT
list or in ORDER BY, and such columns are materialized, should there
be a streaming step below the sort access path.

Change-Id: Iadd9cf7a43f95aaf6a650c1c400166a4adf8f198