Bug#34231798 Incorrect query result from pushed join with IN-subquery
The failing queries are 'semi-joins', which semantically are expected
to eliminate join duplicates after the first match has been found - Contrary
to normal joins where all matching rows should be returned (in any order).

Thus the differens semi-join iterators in the mysql server does some kind
of skip-read after the first matching set of row(s) has been found for a
semi-join nest of tables. This may also skip over result rows from
other tables depending on the table(s) being skip-read. I.e tables
being in the same query tree branch as the rows being skipped.

That is fine when these tables are a part of the same semi-join as
being skip-then - Then this is intended behavior.

However, we sometimes ends up with query plans where the semi-join'ed
tables are evaluated first, and the inner joined tables ends up
depending in the semi-join'ed parts. Usually (only?) seen when the
'duplicate eliminate' iterators are use din the query plan.

Note that this effectively turns the table order in the originating
SQL query upside down. E.g. the pseudo SQL query:

   select ... from t1 where <column1> in (select <column2> from t2 where <pred>)

Might get the query plan

   duplicate eliminate (select <column2> from t2) join t1 on <pred>

Thus, we have a plan where t1 depends on a semi-joined t2, without being
part of the semi-join itself. However it will have t2 as an ancestor in the
SPJ query tree if the query is pushed -> t1 becomes a subject of
the t2 duplicate elimination, effectively a skip-read operation

Due to the finite size of the batch row buffers when returning SPJ results
to the API, we might need to return t1 result rows over multiple batches,
with the t2 result rows being reused/repeated. Thus they will appear as
dupliacted to the iterators, and be skipped over, together with the t1 rows
which should not have been skipped.

Patch identifies when we have such query plans where non-semi-joined
tables are depending on semi-joined tables, _and_ both tables are
scan operation subject to such batching mechanisms. We will then reject
pushing of depending scan-tables not being an intended part of the
semi-joins itself.

Note that such query plans seems to be a rare corner case.

Patch also changes some test cases where:
 - Added two variants of existing test cases where coverage of
   duplicate eliminating iterators were not sufficient
 - Added SEMIJOIN(LOOSESCAN) hint to enisure that intended planes where produced.
 - Added two test cases for bug itself.

That ^ smoked out a query plan which returned incorrect results after modification.
With the patch pushability was reduced, and result became correct.

Change-Id: Iae890ef702cac8a50564d5fb0e493a4715c4dafd