BUG#34555045 - join rejected in group bootstrapped with paxos single leader but runtime value 0
Problem
-------------------
`performance_schema.replication_group_communication_information.
WRITE_CONSENSUS_SINGLE_LEADER_CAPABLE` reflects the runtime value of
Paxos Single Leader setup in a group, and its main purpose is to let
users know which must be the value of
`group_replication_paxos_single_leader` on joining members.

A group that was bootstrapped with single leader enabled but its
protocol version is downgraded to one that does not support it reports
WRITE_CONSENSUS_SINGLE_LEADER_CAPABLE=0, as expected. However,
attempting to join an instance to the group using
group_replication_paxos_single_leader=0 fails

Analysis and Fix
-------------------

For this, we will change the behaviour and make the value of
`group_replication_paxos_single_leader` to be up to par with
the Communication Version that the group is running.

`group_replication_paxos_single_leader` was introduced in 8.0.27,
and below that version, it is not known or used. As such, we will
enforce the following rules:

- When a node joins a group that is running < 8.0.27 and we are of a
  version >= 8.0.27, we must error out and state that
  `group_replication_paxos_single_leader` must be OFF before joining
  the group
- When we try to run `set_communication_protocol` to a version < 8.0.27
  and we are of a version >= 8.0.27, we must error out the UDF if
  `group_replication_paxos_single_leader` is not OFF

This bug also changes the value that we used to check if we are allowed
to change the group leader after running `set_communication_protocol`.
Until today, we would look at the runtime value of
`group_replication_paxos_single_leader`. This is not
correct, since, as per the WL design, this value only takes effect
after a group reboot. As such, when we run `set_communication_protocol`,
we will use the value that is shown in `performance_schema.
replication_group_communication_information
.WRITE_CONSENSUS_SINGLE_LEADER_CAPABLE`

Change-Id: I9a73ec0f58ddfdf36694287a2ece0c2524e2c2da