Bug#34748973 - CPU AFFINITY IS NOT DROPPED ON RESOURCE GROUP DROP
When resource group assigned to a thread is dropped, USR_default
resource group is assigned to it. USR_default resource group
has "0" CPU priority and no CPU affinity. So thread should be
able to use any CPU in this case. But thread is bound to CPU's of
a dropped resource group. This issue is observed on Linux and
FreeBSD only.

While applying default resource group, CPU affinity is dropped
by setting bits for all available CPUs in the system. To get
number of CPUs in Linux and FreeBSD, pthread_getaffinity_np() is
invoked incorrectly. This method actually returns CPU_SET with
bits set for CPUs to which invoking thread is bound. If thread
dropping resource group is assigned to the same resource group
then all threads assigned to resource groups keeps on using same
the set of CPUs instead of using all CPUs.

To fix this issue, changed method to get correct CPU count.

Change-Id: I5def0582ff0a3c07f886b2a18e2c95c8a0c1e2bc