BUG#34240178: gr_member_actions_error_on_read_on_mpm_to_spm failing on PB2
Test gr_member_actions_error_on_read_on_mpm_to_spm does test how
a group mode switch handles a failure during the update of the
member actions table, causing the member to leave the group.
That is achieved by enabling a debug flag that returns a error
when we close the member actions table. Though that flag, which is
set on a common code path can affect other steps of the group mode
switch, which will continue to fail and leave the group.
The test was failing because the expected error message was not
logged into the error log, which means that the group mode switch
did error out before reaching the member actions table error.

Given that the point on which group mode switch fails is not
deterministic, we remove the error log message assert from the test.

Change-Id: I42c9e3564f79c15b80ae99a1c2edee634be0f524