BUG#34527407: gr_parallel_start_uninstall daily-trunk PB2 fail
gr_parallel_start_uninstall forces a race condition between
  `UNINSTALL PLUGIN group_replication;`
and
  `START GROUP_REPLICATION;`
Despite the test first asynchronously executes the `UNINSTALL`,
there is the possibility that the `START` is executed first.
`START` does enabled `super_read_only`, disabling it after the
member joins the group and it is a primary.
When the `UNINSTALL` is allowed to execute once the `START`
is complete, that may happen before the `super_read_only` is
disabled.
If that happens the `UNINSTALL` will hit the error:
  ERROR 1290 (HY000): The MySQL server is running with the
  --super-read-only option so it cannot execute this statement

Since the above error is possible, we added to one of possible error
status of `UNINSTALL PLUGIN group_replication;`.

Change-Id: I9847def076ec1236a2e273befbef52d3fcdf1376