Bug#34291261 The result of query plan is different between  5.7 and 8.0
When optimizer has 2 index choices, one PRIMARY and one SECONDARY to filter
rows, in 5.7 it picks range scan on SECONDARY index because range scan uses
more keyparts. However, 8.0 does not use this logic. Instead, it picks the
PRIMARY index to filter rows with WHERE clause filtering. Primary key use is
not suitable due to limit and the nature of data distribution. Secondary index
is not considered while resolving order by due to the const elimination logic
issue.

Details: Range optimizer picks PRIMARY index over SECONDARY index due to cost
although both could be used (LIMIT consideration and plan update happens later).
At this point, for an index, key_parts are marked as constant, where as the
optimization that removes the constant parts (equalities in where) of the order
by is not yet done. However, the logic in test_if_order_by_key() presumes that
optimization is already done and therefore concludes that SECONDARY index cannot
be used as ordering index. Later in test_if_skip_sort_order(), optimizer tries
to check for an ordering index and it continues with the choice made by range
optimizer because its usable as the ordering key (now that the removal of
constant parts of the ORDER by is done).

Solution: Skip the const key_parts of index during order by evaluation in
test_if_order_by_key() only if the query is const optimized. Query will be const
optimized in test_if_skip_sort_order() but not during LIMIT analysis.

Change-Id: Ideb829424a4ff6df284b3c35cea36de1c06cc34fi