Bug#34768216 SPJ should scan local fragments first
When the SPJ block submits SCAN_FRAGREQs to the LDM's, it will usually scan
only a subsets of the fragments in parallel. Based on recsPrKeys-statistics
if 'valid', or just make a guess of no statistics is available.

SPJ have logic which may take advantage of the result collected from
the first fragments scanned:
 - Parallelism statistics are collected after SCAN_FRAGCONFs
 - firstMatch elimination may eliminate keys needed to scan in
   the next rounds.

As scanning the local fragments are expected to have less overhead than
the non-local fragments, it is preferably to err on the scan-parallelism
for the local fragments. The same should be expected for the 'firstMatch'
as well (when implemented, see bug#34768191)

Patch introduce two rounds to be made over the fragments in ScanFrag_send():
 - first round will only allow sending SCAN_FRAGREQs to local fragments.
 - Second round may send to any fragment expecting a SCAN_FRAGREQ

Change-Id: I0b7e024edd0b909661ca8a4929e33d969c3a5233