Bug#35037016 handle TLS shutdown alert after COM_QUIT
The TLS shutdown Alert message is sent by the TLS protocol when
a TLS session is closed by one of the participants.

It is sent in the MySQL protocol after the COM_QUIT and allows close the
session and reuse for a new connection.

Currently, the shutdown-alert is neither handled, nor sent.

Change
------

- shutdown the SSL session after a COM_QUIT by sending an
  SSL shutdown Alert and the COM_QUIT in the same
  packet.

Change-Id: I4473ce03163fa907b13b95123afeb681c7262924